+++ src/components/class/Chats.jsx
@@ -1,0 +1,5 @@
+const Chats = () => {
+  return <>Chats</>;
+};
+
+export default Chats;
