+++ src/components/class/ClassFrame.jsx
@@ -1,0 +1,11 @@
+import ChatFrame from "./ChatFrame";
+
+const ClassFrame = () => {
+  return (
+    <>
+      <ChatFrame />
+    </>
+  );
+};
+
+export default ClassFrame;
